Add Criterion benchmarking
