+++ Cargo.toml
@@ -24,5 +24,9 @@ rayon = "1.0.3"
 
 [dev-dependencies]
 rand = "0.6.5"
+criterion = "0.2.10"
 
+[[bench]]
+name = "lzjd"
+harness = false
